% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmat.R
\name{set_coords}
\alias{set_coords}
\title{Set coordinates for weights matrix}
\usage{
set_coords(x, coords)
}
\arguments{
\item{x}{a [wmat()] object}

\item{coords}{an object inheriting from "sfc_POINT", see [sf::st_sfc()] and
[sf::st_point()].}
}
\description{
Set coordinates that are used to plot vertices for the weights matrix.
}
\details{
The weights matrices returned by [w_contiguity()] etc. return a
  [wmat()] object without the optional "coords" argument that is used to
  plot the vertices. `set_vertices` allows setting these after the fact,
  and is used for example internally in [wstates()] to set the coords to
  the state capital coordinates.
}
