% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatlag.R
\name{spatlag}
\alias{spatlag}
\title{Spatial lagger}
\usage{
spatlag(x, ccode, date, w_func, ...)
}
\arguments{
\item{x}{variable to lag}

\item{ccode}{vector of numeric G&W country codes to subset}

\item{date}{vector of date for which to construct state set}

\item{w_func}{foo}
}
\description{
Spatial lagger for country-year data
}
