% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states-geom.R
\name{states_geom}
\alias{states_geom}
\title{State borders}
\usage{
states_geom(date, ccode = NULL)
}
\arguments{
\item{date}{A Date or YYYY-MM-DD string}

\item{ccode}{Vector of numeric G&W country codes to subset the result by}
}
\description{
Get correct set of state borders for a date
}
\examples{
x <- states_geom("2010-01-01")
x

}
