% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatlag.R
\name{wstates}
\alias{wstates}
\title{State W}
\usage{
wstates(date, w_func = w_contiguity, ccode = NULL)
}
\arguments{
\item{date}{A Date or YYYY-MM-DD string}

\item{w_func}{type of spatial weighting function}

\item{ccode}{Optional vector of numeric G&W country codes to subset by}
}
\description{
Spatial weights matrix for independent states on a date
}
\examples{
w <- wstates("2010-01-01")
w
plot(w)

# state capitals are by default used for the vertices when plotting
# reset to null to default to centroids
attr(w, "coords") <- NULL
plot(w)

}
