% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmat.R
\name{wmat}
\alias{wmat}
\title{Spatial weights matrix}
\usage{
wmat(w, geom)
}
\arguments{
\item{w}{foo}

\item{geom}{foo}
}
\description{
Spatial weights matrix class / constructor
}
\examples{
data("est_adm1")
w <- w_contiguity(sf::st_geometry(est_adm1))
w
plot(w)

}
